+++ src/commonMain/kotlin/com/github/ageofwar/ktelegram/Audio.kt
@@ -12,6 +12,6 @@ data class Audio(
     val title: String? = null,
     @SerialName("file_name") val fileName: String? = null,
     @SerialName("mime_type") val mimeType: String? = null,
-    @SerialName("file_size") override val fileSize: Int,
+    @SerialName("file_size") override val fileSize: Long,
     @SerialName("thumb") val thumbnail: PhotoSize? = null
 ) : TelegramFile
