+++ src/commonMain/kotlin/com/github/ageofwar/ktelegram/TelegramFile.kt
@@ -2,5 +2,5 @@ package com.github.ageofwar.ktelegram
 
 interface TelegramFile : Id<String> {
     val fileId: String
-    val fileSize: Int
+    val fileSize: Long
 }
